+++ js/ui_components/ui_components.js
@@ -4,3 +4,4 @@ export * from "./ui_base_components.js";
 export * from "./ui_parameters.js";
 export * from "./ui_append_help.js";
 export * from "./ui_object_frame.js";
+export * from "./ui_loading_frame.js";